compiler, runtime: don't let slices point past end of memory block
When Go code uses a slice expression like [:len(str)] or [:cap(slice)],
it's natural for the resulting pointer to point just past the end of
the memory block.  If the next memory block is not live, we now have a
live pointer to a dead block, which will unnecessarily keep the block
alive.  That wastes space, and with the new Go 1.8 GC (not yet
committed) will trigger an error when using GODEBUG=gccheckmark=1.

This changes the implementation of slice expressions to not move the
pointer if the resulting string length or slice capacity is 0.  When
the length/capacity is zero, the pointer is never used anyhow.

Change-Id: I4902c3f26f5700c4a1c84d2efdd8decc4df122d4
Reviewed-on: https://go-review.googlesource.com/39870
Reviewed-by: Than McIntosh <thanm@google.com>