compiler: add code to generate a ptrmask for a type
The Go 1.8 garbage collector uses a ptrmask for all types below a
certain size.  A ptrmask is simply a bit vector with a single bit for
each pointer-sized word in the value.  The bit is 1 if the type has a
pointer in that position, 0 if it does not.

This change adds code to the compiler to generate a ptrmask.  The code
is not used by anything yet, it is just compiled.  It will be used
when we switch over to the Go 1.8 garbage collector.

The new Array_type::int_length method, and the new memory_size
methods, will also be used by other patches coming later.

Change-Id: I0b97a3845712fe63fb79efeedae6a0c1d476c306
Reviewed-on: https://go-review.googlesource.com/39873
Reviewed-by: Than McIntosh <thanm@google.com>