compiler: add code to generate a gcprog for a type
The Go 1.8 garbage collector uses a gcprog for all types above a
certain size.  A gcprog describes where the pointers are in the type,
using a simple bytecode machine that supports repeating bits.  The
effect is to permit using much less space to describe arrays.  The
format is described in runtime/mbitmap.go in the docs for runGCProg.
This is not yet added to the gofrontend, but can be seen in the gc sources.

This change adds code to the compiler to generate a gcprog.  The code
is not used by anything yet, it is just compiled.  It will be used
when we switch over to the Go 1.8 garbage collector.

Change-Id: I73882808afea7b35c2156e342c873d957d9185d7
Reviewed-on: https://go-review.googlesource.com/39923
Reviewed-by: Than McIntosh <thanm@google.com>