compiler: add ptrdata computations and expressions
For the upcoming Go 1.8 GC we need to compute the "ptrdata" of a type:
the number of bytes in the type that can contain pointers.  For types
that do not contain pointers this number is zero.  For many types it
is a number larger than zero but smaller than the total size of the
type.  The Go 1.8 GC uses this number to make loops looking for
pointers run faster by not scanning the suffix of a value that can not
contain a pointer.

Unfortunately there are two subtly different definitions of ptrdata,
and we need both.  The first is the simple one: the prefix that can
contain pointers.  The second is the number of bytes described by the
gcprog for the type.  Recall that we describe the exact position of
pointers in a type using either a ptrmask or a gcprog.  The ptrmask is
simpler, the gcprog uses less space.  We use the gcprog for large
types, currently defined as types that are more than 2048 bytes.  When
the Go 1.8 runtime expands a gcprog, it verifies that the gcprog
describes exactly the same number of bytes as the ptrdata field in the
type descriptor.  If the last pointer-containing portion of a type is
an array, and if the elements of the array have a ptrdata that is less
than the size of the element type, then the simple definition of the
ptrdata will not include the final non-pointer-containing bytes of the
last element of the array.  However, the gcprog will define the array
using a repeat count, and will therefore include the full size of the
last element of the array.  So for a type that needs a gcprog, the
ptrdata field in the type descriptor must be the size of the data
described by the gcprog, and that is not necessarily the same as the
simple ptrdata.

It might seem that we can always use the gcprog version of the ptrdata
calculation, since that is what will appear in a type descriptor, but
it turns out that for global variables we always use a ptrmask, not a
gcprog, even if the global variable is large.  This is because gcprogs
are handled by expanding them into a ptrmask at runtime, and for a
global variable there is no natural place to put the ptrmask.  Simpler
to always use the ptrmask.  That means that we need to describe the
size of the ptrmask, and that means that we need an expression for the
simple form of the ptrdata.

This CL implements the ptrdata calculation.  This code is not actually
used yet.  It will be used later when the Go 1.8 GC is committed.

Change-Id: I695e5284d888b95fe71e802f0ad1c9f8f8c2226f
Reviewed-on: https://go-review.googlesource.com/40573
Reviewed-by: Than McIntosh <thanm@google.com>