compiler: don't write struct with multiple sink fields to C header file
When writing a struct to the C header file used by the C runtime code,
a single sink field is fine: it will be called "_", which is valid C.
There are structs with single sink fields that we want to write out,
such as finblock.  As it happens, though, the Go 1.8 runtime has a
struct with two sink fields, gcControllerState, which will produce a C
definition with two fields named "_", which will fail.  Since we don't
need to know that struct in C, rather than fix the general case, just
punt if the struct has multiple sink fields.

After the conversion to the Go 1.8 GC, we may be able to get rid of
the C header file anyhow.  I'm not sure yet.

Change-Id: I88669f015994fe2d67eb85b874b328596e2eb4c8
Reviewed-on: https://go-review.googlesource.com/40701
Reviewed-by: Than McIntosh <thanm@google.com>