libgo: delete goc2c
The last .goc file has been removed, so remove goc2c.

The goc2c program was my first contribution to the gc repository that
was more than 100 lines:
https://github.com/golang/go/commit/2b57a1124e87b0dc8bc1ff6899297b4d7d6e74f2
The program was used in gc for a few years under various guises but
was finally removed in https://golang.org/cl/132680043.  Now we can
remove it from gofrontend as well.

Change-Id: If4e5a617d8ae98697223ea1e8c42efb8826114ab
Reviewed-on: https://go-review.googlesource.com/42911
Reviewed-by: Than McIntosh <thanm@google.com>