mksigtab.sh: recurse once when adding signals to SIGLIST
On MIPS, SIGABRT is defined like this:
 #define SIGIOT   6
 #define SIGABRT  SIGIOT

This breaks addsig which tries to append __SIGIOT_ to SIGLIST. Signal
number 6 is later added to the output and go complains about a
duplicate signal number.

Fix by recursing once when obtaining the signal number from
gen-sysinfo.go if the signal is defined as an alias of another signal.
Also modify the sed expression to 's/.* = //' which is equivalent to
the original expression but is less misleading given that it might not
match a number.

Change-Id: I1a1e547c0af8ff93e906f6bbaa8f1a5704a1dfaf
Reviewed-on: https://go-review.googlesource.com/43252
Reviewed-by: Ian Lance Taylor <iant@golang.org>