libgo: add "vendor" to pkgpath for vendored standard packages
Ensure that the packages vendored into the standard library do not
have the same pkgpath as the actual packages.  If we don't, attempts
to build and test the actual packages will get confused.  The specific
error I was seeing was import loops, causing some of the packages to
fail to get initialized, causing an obscure run time crash.

Change-Id: I399b882a9f3269f6ba6ce024dfb58e0a11301fd5
Reviewed-on: https://go-review.googlesource.com/43610
Reviewed-by: Than McIntosh <thanm@google.com>