compiler: ignore struct field tags for type conversion
Go 1.8 includes a language change (https://golang.org/doc/go1.8#language):
in an explicit conversion from one struct type to another, any field
tags are ignored.

This CL implements this language change in the gofrontend.  The tests
for this are in the gc testsuite, which will be copied into the gccgo
repository in due course.

Updates golang/go#16085.

Change-Id: I64e49f103efced41a55757e2cb665f6c1098db80
Reviewed-on: https://go-review.googlesource.com/43614
Reviewed-by: Than McIntosh <thanm@google.com>