runtime: ignore _Gscan bit when checking status in CgocallDone
Also always access the atomicstatus field atomically.

The effect of not checking the _Gscan bit is that if the GC decides to
scan the stack just as the goroutine is leaving the system call, the
goroutine might fail to call exitsyscall.  Then then typically causes
a runtime assertion failure later on.  If we do call exitsyscall as we
should, it will stall (in casgstatus) until the _Gscan bit is cleared.

No separate test.  I've observed causing sporadic failures running the
misc/cgo tests, but we don't currently have a way to run those
routinely for gccgo.  I should fix that.

Change-Id: Id1ae47bb5c619e8b299b2fce53717e663c84edf8
Reviewed-on: https://go-review.googlesource.com/45392
Reviewed-by: Than McIntosh <thanm@google.com>