internal/testenv: support testing gccgo
If GO_TESTING_GOTOOLS is set in the environment, permit tests using
gccgo to run the go tool. Like GO_BUILDER_NAME, this should not be set
normally. But it is needed when testing the go tool itself, and will
be set by the gotools Makefile in a future CL.

Change-Id: Ifce3185ff03bd966fbc26ae2adbbf2777231630b
Reviewed-on: https://go-review.googlesource.com/45693
Reviewed-by: Than McIntosh <thanm@google.com>