compiler: avoid crash on erroneous type
If there is an error constructing the backend type, the GCC backend
will report that the size is 1.  That will then cause construction of
the ptrmask to crash.  Avoid that case by just generating an empty
ptrmask.

Noticed while compiling a broken package.  The policy I've been
following is to not commit a test case for a compiler crash on invalid
code, so no test case.

Change-Id: Ied85154d9d49012be9d693c4488a077c214efb7e
Reviewed-on: https://go-review.googlesource.com/45775
Reviewed-by: Than McIntosh <thanm@google.com>