runtime: don't build test files that use SetCgoTraceback for gccgo
The gofrontend doesn't support the runtime.SetCgoTraceback function,
which is specifically for handling mixed Go and C tracebacks.
Use a build tag to avoid compiling the runtime/testdata/testprogcgo
files that refer to SetCgoTraceback. These files are not currently
compiled anyhow, but they will be with a future gotools patch.

Change-Id: I016c5eb21503c49aefa16727b023c7381f2c0d18
Reviewed-on: https://go-review.googlesource.com/46452
Reviewed-by: Than McIntosh <thanm@google.com>