runtime: avoid write barriers when calling deferred function
Calling a deferred function currently requires changing from a uintptr
to the function code to a Go function value. That is done by setting
the value of a func local variable using unsafe.Pointer. The local
variable will always be on the stack. Adjust the code that sets the
local variable to avoid generating a write barrier.

A write barrier is never needed here. Also, for deferreturn, we must
avoid write barriers entirely when called from a cgo function; that
requires more than just this, but this is a start.

The test for this is runtime tests that use the go tool; these are not
currently run, but they will be in the future.

Change-Id: I762fd07b7fd5b46c0c54156f89c71fb60de53023
Reviewed-on: https://go-review.googlesource.com/46455
Reviewed-by: Than McIntosh <thanm@google.com>