runtime: don't crash if no p in kickoff
The kickoff function for g0 can be invoked without a p, for example
from mcall(exitsyscall0) in exitsyscall after exitsyscall has cleared
the p field. The assignment gp.param = nil will invoke a write barrier.
If gp.param is not already nil, this will require a p. Avoid the problem
for a specific case that is known to be OK: when the value in gp.param
is a *g.

Change-Id: Ifc0451359a2b09a62bff51159186a10900bd9f01
Reviewed-on: https://go-review.googlesource.com/46512
Reviewed-by: Than McIntosh <thanm@google.com>