compiler: centralize all symbol name handling
Consolidate all symbol name handling into the new file names.cc.  This
is intended to define all the names that will appear in the generated
assembly code.  Names that will not appear in the assembly code, such
as local variable names or label names, remain where they are.

This consolidation is not intended to change any of the existing
symbol names.  Tested by building without and without this patch and
comparing the libgo symbol table.

Change-Id: Ibeab4591d0648fa12def570092e06316c7cf2b9b
Reviewed-on: https://go-review.googlesource.com/68310
Reviewed-by: Than McIntosh <thanm@google.com>