compiler: add backend type conversion
Tweak Interface_field_reference_expression::do_get_backend to apply an
additional backend type conversion to the returned result.  This is
needed due to the fact that the top level type of the expression is a
function descriptor, however the value being manufactured is a pointer
to <thunk-descriptor, value> struct.

Change-Id: I435c02d43f76e179bec8e95e5dd0ea9f47a62eae
Reviewed-on: https://go-review.googlesource.com/51712
Reviewed-by: Ian Lance Taylor <iant@golang.org>