compiler: fix check for notinheap conversion
A normal pointer may not be converted to a notinheap pointer.  We were
erroneously permitting a conversion from a normal pointer to a
notinheap unsafe.Pointer, which is useless since unsafe.Pointer is not
marked notinheap.  Correct the test to permit a conversion from
unsafe.Pointer to a notinheap pointer, which is the same test that the
gc compiler uses.

The test case for this is in the 1.9 runtime package.

Change-Id: Ie86db54020ed977068969d53fa67568c3cca8d23
Reviewed-on: https://go-review.googlesource.com/62731
Reviewed-by: Than McIntosh <thanm@google.com>