runtime: export cgoCheck functions
The functions cgoCheckPointer and cgoCheckResult are called by code
generated by cgo. That means that we need to export them using
go:linkname, as otherwise they are local symbols. The cgo code
currently uses weak references to only call the symbols if they are
defined, which is why it has been working--the cgo code has not been
doing any checks.

Change-Id: I872c429fef02a9ba0eddc25f72e84e346866a0a3
Reviewed-on: https://go-review.googlesource.com/80295
Reviewed-by: Than McIntosh <thanm@google.com>