compiler: move some escape check to Mark_address_taken
Move some check of escape state earlier, from get_backend to
Mark_address_taken. So we can reclaim escape analysis Nodes
before kicking off the backend (not done in this CL). Also it
makes it easier to check variables and closures do not escape
when the escape analysis is run for the runtime package (also
not done in this CL).

Change-Id: I55c37e688b46e661b68ebc15aac7a045600086e9
Reviewed-on: https://go-review.googlesource.com/85735
Reviewed-by: Ian Lance Taylor <iant@golang.org>